nix!: Remove TEX & NerdFonts 79 deps
This was a major refactor, and improved build speed by a significant
margin.
On top of the cleanup mentioned in the commit title Postman & checkov
were moved back to the `brew` flow, as they were having invalid Wayland
dependencies, which was causing release to fail.

nix: d2 still the same

nix: surp

sdj